--Added savings of all user`s ratings in memory. --Changed rating type to byte

+++ HiP-DataStore.Model/Rest/RatingArgs.cs
@@ -7,7 +7,7 @@ namespace PaderbornUniversity.SILab.Hip.DataStore.Model.Rest
     {
         [Required]
         [Range(1, 5, ErrorMessage = "Has to be in a range from 1 to 5")]
-        public int? Rating { get; set; }
+        public byte? Rating { get; set; }
 
         [Required]
         [Range(0, Int32.MaxValue, ErrorMessage = "Has to be a positive value")]
